feat: added feature flag for sending events on ScUsers topic (#168)
